Location tracking now works
Using low-power location tracking (only invokes GPS for the first time
and when a zone is exited). For the most part, worked out the logic
needed to keep current location view in sync with the switch on the
flip side (a bug remains when a page is removed, resulting in crash,
this needs to be solved...).