Fixed timer problems in MainViewController
Need to retain timer object, otherwise calling [timer invalidate] will
cause app to crash. For now, placing timer creation and destruction
into viewDidAppear and viewDidDisappear respectively (the problem with
that is that timer gets invalidated every time fipside view is shown,
which is not a huge deal but something that should be best avoided).